+++ src/main/java/com/fernando9825/alcaldiasvrestapi/models/services/IMunicipioService.java
@@ -13,5 +13,5 @@ public interface IMunicipioService {
 
     Institucion save(Municipio municipio);
 
-    void delete(Long municipioId);
+    void delete(Integer municipioId);
 }
